# Conflicts:
#	src/main/java/com/sparta/projectblue/domain/performance/dto/PerformanceDetailDto.java

+++ src/main/java/com/sparta/projectblue/domain/performance/dto/PerformanceDetailDto.java
@@ -13,7 +13,7 @@ public class PerformanceDetailDto {
     private String title;
     private LocalDateTime startDate;
     private LocalDateTime endDate;
-    private int price;
+    private Long price;
     private String category;
     private String description;
     private int duration;
@@ -22,7 +22,7 @@ public class PerformanceDetailDto {
     private String imageUrl;
 
     // 지우지 마세요. 쿼리에서 써요.
-    public PerformanceDetailDto(String title, LocalDateTime startDate, LocalDateTime endDate, int price, String category, String description, int duration, String hallName,String posterName, String imageUrl) {
+    public PerformanceDetailDto(String title, LocalDateTime startDate, LocalDateTime endDate, Long price, String category, String description, int duration, String hallName,String posterName, String imageUrl) {
         this.title = title;
         this.startDate = startDate;
         this.endDate = endDate;
